update smaugt and benchmarks

+++ crypto_kem/smaugt5/m4/hash.c
@@ -1,8 +1,8 @@
 #include "hash.h"
 
-void shake256_absorb_twice_squeeze(uint8_t *out, unsigned long long out_bytes,
-                                   const uint8_t *in1, unsigned long long in1_bytes,
-                                   const uint8_t *in2, unsigned long long in2_bytes) {
+void shake256_absorb_twice_squeeze(uint8_t *out, size_t out_bytes,
+                                   const uint8_t *in1, size_t in1_bytes,
+                                   const uint8_t *in2, size_t in2_bytes) {
     shake256incctx state;
     shake256_init(&state);
     shake256_absorb(&state, in1, in1_bytes);
